requirements.txt
fastapi
uvicorn
openai
python-docx
PyPDF2
fpdf
